python/requirements.txt
langchain==0.0.339
octoai-sdk==0.7.0
openai==0.28.1